+++ tests/machine.test.ts
@@ -59,40 +59,40 @@ describe(OlmMachine.name, () => {
     });
 
     test("can be instantiated with a store", async () => {
-        let store_name = "hello";
-        let store_passphrase = "world";
+        let storeName = "hello";
+        let storePassphrase = "world";
 
-        const by_store_name = (db: IDBDatabaseInfo) => db.name!.startsWith(store_name);
+        const byStoreName = (db: IDBDatabaseInfo) => db.name!.startsWith(storeName);
 
         // No databases.
-        expect((await indexedDB.databases()).filter(by_store_name)).toHaveLength(0);
+        expect((await indexedDB.databases()).filter(byStoreName)).toHaveLength(0);
 
         // Creating a new Olm machine.
         expect(
-            await OlmMachine.initialize(new UserId("@foo:bar.org"), new DeviceId("baz"), store_name, store_passphrase),
+            await OlmMachine.initialize(new UserId("@foo:bar.org"), new DeviceId("baz"), storeName, storePassphrase),
         ).toBeInstanceOf(OlmMachine);
 
-        // Oh, there is 2 databases now, prefixed by `store_name`.
-        let databases = (await indexedDB.databases()).filter(by_store_name);
+        // Oh, there is 2 databases now, prefixed by `storeName`.
+        let databases = (await indexedDB.databases()).filter(byStoreName);
 
         expect(databases).toHaveLength(2);
         expect(databases).toStrictEqual([
-            { name: `${store_name}::matrix-sdk-crypto-meta`, version: 1 },
-            { name: `${store_name}::matrix-sdk-crypto`, version: 7 },
+            { name: `${storeName}::matrix-sdk-crypto-meta`, version: 1 },
+            { name: `${storeName}::matrix-sdk-crypto`, version: 7 },
         ]);
 
         // Creating a new Olm machine, with the stored state.
         expect(
-            await OlmMachine.initialize(new UserId("@foo:bar.org"), new DeviceId("baz"), store_name, store_passphrase),
+            await OlmMachine.initialize(new UserId("@foo:bar.org"), new DeviceId("baz"), storeName, storePassphrase),
         ).toBeInstanceOf(OlmMachine);
 
         // Same number of databases.
-        expect((await indexedDB.databases()).filter(by_store_name)).toHaveLength(2);
+        expect((await indexedDB.databases()).filter(byStoreName)).toHaveLength(2);
     });
 
     describe("cannot be instantiated with a store", () => {
         test("store name is missing", async () => {
-            let store_passphrase = "world";
+            let storePassphrase = "world";
 
             let err = null;
 
@@ -101,7 +101,7 @@ describe(OlmMachine.name, () => {
                     new UserId("@foo:bar.org"),
                     new DeviceId("baz"),
                     undefined,
-                    store_passphrase,
+                    storePassphrase,
                 );
             } catch (error) {
                 err = error;
@@ -111,12 +111,12 @@ describe(OlmMachine.name, () => {
         });
 
         test("store passphrase is missing", async () => {
-            let store_name = "hello";
+            let storeName = "hello";
 
             let err = null;
 
             try {
-                await OlmMachine.initialize(new UserId("@foo:bar.org"), new DeviceId("baz"), store_name, undefined);
+                await OlmMachine.initialize(new UserId("@foo:bar.org"), new DeviceId("baz"), storeName, undefined);
             } catch (error) {
                 err = error;
             }
@@ -129,10 +129,10 @@ describe(OlmMachine.name, () => {
     const device = new DeviceId("foobar");
     const room = new RoomId("!baz:matrix.org");
 
-    function machine(new_user?: UserId, new_device?: DeviceId): Promise<OlmMachine> {
+    function machine(newUser?: UserId, newDevice?: DeviceId): Promise<OlmMachine> {
         // Uncomment to enable debug logging for tests
         // new RustSdkCryptoJs.Tracing(RustSdkCryptoJs.LoggerLevel.Trace).turnOn();
-        return OlmMachine.initialize(new_user || user, new_device || device);
+        return OlmMachine.initialize(newUser || user, newDevice || device);
     }
 
     test("can drop/close", async () => {
@@ -141,38 +141,38 @@ describe(OlmMachine.name, () => {
     });
 
     test("can drop/close with a store", async () => {
-        let store_name = "temporary";
-        let store_passphrase = "temporary";
+        let storeName = "temporary";
+        let storePassphrase = "temporary";
 
-        const by_store_name = (db: IDBDatabaseInfo) => db.name?.startsWith(store_name);
+        const byStoreName = (db: IDBDatabaseInfo) => db.name?.startsWith(storeName);
 
         // No databases.
-        expect((await indexedDB.databases()).filter(by_store_name)).toHaveLength(0);
+        expect((await indexedDB.databases()).filter(byStoreName)).toHaveLength(0);
 
         // Creating a new Olm machine.
         const m = await OlmMachine.initialize(
             new UserId("@foo:bar.org"),
             new DeviceId("baz"),
-            store_name,
-            store_passphrase,
+            storeName,
+            storePassphrase,
         );
         expect(m).toBeInstanceOf(OlmMachine);
 
-        // Oh, there is 2 databases now, prefixed by `store_name`.
-        let databases = (await indexedDB.databases()).filter(by_store_name);
+        // Oh, there is 2 databases now, prefixed by `storeName`.
+        let databases = (await indexedDB.databases()).filter(byStoreName);
 
         expect(databases).toHaveLength(2);
         expect(databases).toStrictEqual([
-            { name: `${store_name}::matrix-sdk-crypto-meta`, version: 1 },
-            { name: `${store_name}::matrix-sdk-crypto`, version: 7 },
+            { name: `${storeName}::matrix-sdk-crypto-meta`, version: 1 },
+            { name: `${storeName}::matrix-sdk-crypto`, version: 7 },
         ]);
 
         // Let's force to close the `OlmMachine`.
         m.close();
 
         // Now we can delete the databases!
-        for (const database_name of [`${store_name}::matrix-sdk-crypto`, `${store_name}::matrix-sdk-crypto-meta`]) {
-            const deleting = indexedDB.deleteDatabase(database_name);
+        for (const databaseName of [`${storeName}::matrix-sdk-crypto`, `${storeName}::matrix-sdk-crypto-meta`]) {
+            const deleting = indexedDB.deleteDatabase(databaseName);
             deleting.onsuccess = () => {};
             deleting.onerror = () => {
                 throw new Error("failed to remove the database (error)");
@@ -335,13 +335,13 @@ describe(OlmMachine.name, () => {
                 expect(request).toBeInstanceOf(KeysUploadRequest);
 
                 // https://spec.matrix.org/v1.2/client-server-api/#post_matrixclientv3keysupload
-                const hypothetical_response = JSON.stringify({
+                const hypotheticalResponse = JSON.stringify({
                     one_time_key_counts: {
                         curve25519: 10,
                         signed_curve25519: 20,
                     },
                 });
-                const marked = await m.markRequestAsSent(request.id!, request.type, hypothetical_response);
+                const marked = await m.markRequestAsSent(request.id!, request.type, hypotheticalResponse);
                 expect(marked).toStrictEqual(true);
             }
 
@@ -350,7 +350,7 @@ describe(OlmMachine.name, () => {
                 expect(request).toBeInstanceOf(KeysQueryRequest);
 
                 // https://spec.matrix.org/v1.2/client-server-api/#post_matrixclientv3keysquery
-                const hypothetical_response = JSON.stringify({
+                const hypotheticalResponse = JSON.stringify({
                     device_keys: {
                         "@alice:example.org": {
                             JLAFKJWSCS: {
@@ -375,7 +375,7 @@ describe(OlmMachine.name, () => {
                     },
                     failures: {},
                 });
-                const marked = await m.markRequestAsSent(request.id!, request.type, hypothetical_response);
+                const marked = await m.markRequestAsSent(request.id!, request.type, hypotheticalResponse);
                 expect(marked).toStrictEqual(true);
             }
         });
@@ -394,7 +394,7 @@ describe(OlmMachine.name, () => {
         test("can pass keysquery and keysclaim requests directly", async () => {
             {
                 // derived from https://github.com/matrix-org/matrix-rust-sdk/blob/7f49618d350fab66b7e1dc4eaf64ec25ceafd658/benchmarks/benches/crypto_bench/keys_query.json
-                const hypothetical_response = JSON.stringify({
+                const hypotheticalResponse = JSON.stringify({
                     device_keys: {
                         "@example:localhost": {
                             AFGUOBTZWM: {
@@ -467,12 +467,12 @@ describe(OlmMachine.name, () => {
                         },
                     },
                 });
-                const marked = await m.markRequestAsSent("foo", RequestType.KeysQuery, hypothetical_response);
+                const marked = await m.markRequestAsSent("foo", RequestType.KeysQuery, hypotheticalResponse);
             }
 
             {
                 // derived from https://github.com/matrix-org/matrix-rust-sdk/blob/7f49618d350fab66b7e1dc4eaf64ec25ceafd658/benchmarks/benches/crypto_bench/keys_claim.json
-                const hypothetical_response = JSON.stringify({
+                const hypotheticalResponse = JSON.stringify({
                     one_time_keys: {
                         "@example:localhost": {
                             AFGUOBTZWM: {
@@ -490,14 +490,14 @@ describe(OlmMachine.name, () => {
                     },
                     failures: {},
                 });
-                const marked = await m.markRequestAsSent("bar", RequestType.KeysClaim, hypothetical_response);
+                const marked = await m.markRequestAsSent("bar", RequestType.KeysClaim, hypotheticalResponse);
             }
         });
 
         test("can share a room key", async () => {
-            const other_users = [new UserId("@example:localhost")];
+            const otherUsers = [new UserId("@example:localhost")];
 
-            const requests = await m.shareRoomKey(room, other_users, new EncryptionSettings());
+            const requests = await m.shareRoomKey(room, otherUsers, new EncryptionSettings());
 
             expect(requests).toHaveLength(1);
             expect(requests[0]).toBeInstanceOf(ToDeviceRequest);
@@ -510,7 +510,7 @@ describe(OlmMachine.name, () => {
             expect(messageContent["org.matrix.msgid"]).toBeDefined();
 
             await m.markRequestAsSent(requests[0].id, RequestType.ToDevice, "{}");
-            const requestsAfterMarkedAsSent = await m.shareRoomKey(room, other_users, new EncryptionSettings());
+            const requestsAfterMarkedAsSent = await m.shareRoomKey(room, otherUsers, new EncryptionSettings());
             expect(requestsAfterMarkedAsSent).toHaveLength(0);
         });
 
@@ -746,7 +746,7 @@ describe(OlmMachine.name, () => {
             let encryptedExportedRoomKeys = OlmMachine.encryptExportedRoomKeys(
                 exportedRoomKeys,
                 encryptionPassphrase,
-                100_000,
+                100000,
             );
 
             expect(encryptedExportedRoomKeys).toMatch(/^-----BEGIN MEGOLM SESSION DATA-----/);
@@ -819,7 +819,7 @@ describe(OlmMachine.name, () => {
 
         test("can inject devices from someone else", async () => {
             {
-                const hypothetical_response = JSON.stringify({
+                const hypotheticalResponse = JSON.stringify({
                     device_keys: {
                         "@example:morpheus.localhost": {
                             ATRLDCRXAC: {
@@ -930,7 +930,7 @@ describe(OlmMachine.name, () => {
                         },
                     },
                 });
-                const marked = await m.markRequestAsSent("foo", RequestType.KeysQuery, hypothetical_response);
+                const marked = await m.markRequestAsSent("foo", RequestType.KeysQuery, hypotheticalResponse);
             }
         });
 
@@ -1024,11 +1024,11 @@ describe(OlmMachine.name, () => {
 
             let keyBackupKey = BackupDecryptionKey.createRandomKey();
 
-            let auth_data = {
+            let authData = {
                 public_key: keyBackupKey.megolmV1PublicKey.publicKeyBase64,
             };
 
-            let canonical = JSON.stringify(auth_data);
+            let canonical = JSON.stringify(authData);
 
             let signaturesJSON = (await m.sign(canonical)).asJSON();
 
@@ -1036,7 +1036,7 @@ describe(OlmMachine.name, () => {
                 algorithm: keyBackupKey.megolmV1PublicKey.algorithm,
                 auth_data: {
                     signatures: JSON.parse(signaturesJSON),
-                    ...auth_data,
+                    ...authData,
                 },
             };
 
@@ -1077,11 +1077,11 @@ describe(OlmMachine.name, () => {
 
             let sessions = exportedKey.rooms["!baz:matrix.org"].sessions;
             // @ts-ignore "object is of type 'unknown'"
-            let session_data = Object.values(sessions)[0].session_data;
+            let sessionData = Object.values(sessions)[0].session_data;
 
             // should decrypt with the created key
             let decrypted = JSON.parse(
-                keyBackupKey.decryptV1(session_data.ephemeral, session_data.mac, session_data.ciphertext),
+                keyBackupKey.decryptV1(sessionData.ephemeral, sessionData.mac, sessionData.ciphertext),
             );
             expect(decrypted.algorithm).toStrictEqual("m.megolm.v1.aes-sha2");
 
@@ -1153,6 +1153,176 @@ describe(OlmMachine.name, () => {
 
             expect(progressListener).toHaveBeenCalledTimes(1);
             expect(progressListener).toHaveBeenCalledWith(0, 1);
+        });
+    });
+
+    describe("Request missing secrets", () => {
+        /**
+         * Creates an hypothetical response to a key query request for an account with a pre-existing device and identity.
+         *
+         * To be used in tests when you want to create a setup where there is an existing device
+         * on the account with cross-signing set up.
+         *
+         * This will create a valid response to a key query request with all needed signatures.
+         *
+         * @param {UserId} userId - The user id
+         * @param {DeviceId} deviceId - The device id
+         *
+         * @returns A valid response to a key query request that can be feed in a second login for that account.
+         */
+        async function getKeyQueryResponseWithExistingDevice(userId, deviceId) {
+            let initialMachine = await OlmMachine.initialize(userId, deviceId);
+            const userIdStr = initialMachine.userId.toString();
+            const deviceIdStr = initialMachine.deviceId.toString();
+
+            let deviceKeys;
+            let outgoingRequests = await initialMachine.outgoingRequests();
+            for (const request of outgoingRequests) {
+                if (request instanceof KeysUploadRequest) {
+                    deviceKeys = JSON.parse(request.body);
+                }
+            }
+            delete deviceKeys.one_time_keys;
+            delete deviceKeys.fallback_keys;
+
+            let bootstrapRequest = await initialMachine.bootstrapCrossSigning(true);
+
+            const crossSigning = JSON.parse(bootstrapRequest.uploadSigningKeysRequest.body);
+            const newSignature = JSON.parse(bootstrapRequest.uploadSignaturesRequest.body);
+
+            const allSignatures = {
+                [userIdStr]: {
+                    ...deviceKeys.device_keys.signatures[userId],
+                    ...newSignature[userIdStr][deviceIdStr].signatures[userId],
+                },
+            };
+
+            deviceKeys.device_keys.signatures = allSignatures;
+
+            return {
+                device_keys: {
+                    [userIdStr]: {
+                        [deviceIdStr]: deviceKeys,
+                    },
+                },
+                ...crossSigning,
+            };
+        }
+
+        /**
+         * This function is designed to work with `getKeyQueryResponseWithExistingDevice` to simulate a scenario where
+         * an existing device (and cross-signing identity), is already associated with the account.
+         *
+         * It will create a new login and process a hypothetical response that includes the existing identity and devices.
+         *
+         * @param {UserId} userId - the user id of the account
+         * @param {DeviceId} deviceId - the id of the new device
+         * @param {*} hypotheticalResponse - the response to the key query request generated by `getKeyQueryResponseWithExistingDevice`
+         *
+         * @returns an olm machine with the new device and the hypothetical response processed.
+         */
+        async function getSecondMachine(userId, deviceId, hypotheticalResponse) {
+            const secondMachine = await OlmMachine.initialize(userId, deviceId);
+
+            const toDeviceEvents = JSON.stringify([]);
+            const changedDevices = new DeviceLists();
+            const oneTimeKeyCounts = new Map();
+            const unusedFallbackKeys = new Set();
+
+            await secondMachine.receiveSyncChanges(
+                toDeviceEvents,
+                changedDevices,
+                oneTimeKeyCounts,
+                unusedFallbackKeys,
+            );
+            let outgoingRequests = await secondMachine.outgoingRequests();
+
+            const request = outgoingRequests[1];
+            expect(request).toBeInstanceOf(KeysQueryRequest);
+
+            await secondMachine.markRequestAsSent(
+                request.id,
+                RequestType.KeysQuery,
+                JSON.stringify(hypotheticalResponse),
+            );
+
+            return secondMachine;
+        }
+
+        test("Should request cross-signing keys if missing", async () => {
+            const userId = new UserId("@alice:example.org");
+            const firstDevice = new DeviceId("ABCDEF");
+            const hypotheticalResponse = await getKeyQueryResponseWithExistingDevice(userId, firstDevice);
+
+            const secondDeviceId = new DeviceId("GHIJKL");
+
+            const secondMachine = await getSecondMachine(userId, secondDeviceId, hypotheticalResponse);
+            const hasMissingSecrets = await secondMachine.requestMissingSecretsIfNeeded();
+
+            expect(hasMissingSecrets).toStrictEqual(true);
+
+            let outgoingRequests = await secondMachine.outgoingRequests();
+
+            let mskRequested = false;
+            let sskRequested = false;
+            let uskRequested = false;
+            for (const request of outgoingRequests) {
+                if (request instanceof ToDeviceRequest) {
+                    const parsed = JSON.parse(request.body);
+                    const message = parsed.messages[userId]["*"];
+                    if (message && message.action === "request") {
+                        if (message.name === "m.cross_signing.master") {
+                            mskRequested = true;
+                        } else if (message.name === "m.cross_signing.self_signing") {
+                            sskRequested = true;
+                        } else if (message.name === "m.cross_signing.user_signing") {
+                            uskRequested = true;
+                        }
+                    }
+                }
+            }
+
+            expect(mskRequested).toStrictEqual(true);
+            expect(sskRequested).toStrictEqual(true);
+            expect(uskRequested).toStrictEqual(true);
+        });
+
+        test("Should not request if there are requests already in flight", async () => {
+            const userId = new UserId("@alice:example.org");
+            const firstDevice = new DeviceId("ABCDEF");
+            const hypotheticalResponse = await getKeyQueryResponseWithExistingDevice(userId, firstDevice);
+
+            const secondDeviceId = new DeviceId("GHIJKL");
+
+            const secondMachine = await getSecondMachine(userId, secondDeviceId, hypotheticalResponse);
+
+            const hasMissingSecrets = await secondMachine.requestMissingSecretsIfNeeded();
+            expect(hasMissingSecrets).toStrictEqual(true);
+
+            const hasMissingSecretsSecondTry = await secondMachine.requestMissingSecretsIfNeeded();
+            expect(hasMissingSecretsSecondTry).toStrictEqual(false);
+        });
+
+        test("Should not request cross signing secrets if known", async () => {
+            const userId = new UserId("@alice:example.org");
+            const firstDevice = new DeviceId("ABCDEF");
+
+            let initialMachine = await OlmMachine.initialize(userId, firstDevice);
+            await initialMachine.bootstrapCrossSigning(true);
+            let keyBackupKey = BackupDecryptionKey.createRandomKey();
+
+            await initialMachine.saveBackupDecryptionKey(keyBackupKey, "3");
+
+            const hasMissingSecrets = await initialMachine.requestMissingSecretsIfNeeded();
+
+            expect(hasMissingSecrets).toStrictEqual(false);
+
+            let outgoingRequests = await initialMachine.outgoingRequests();
+
+            let toDeviceRequests = outgoingRequests.filter((request) => {
+                return request instanceof ToDeviceRequest;
+            });
+            expect(toDeviceRequests).toHaveLength(0);
         });
     });
 });